% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditiveLogLikelihood.R
\name{AdditiveLogLikelihood}
\alias{AdditiveLogLikelihood}
\title{Additive Model Log Likelihood}
\usage{
AdditiveLogLikelihood(
  param,
  f.hat,
  w0.hat.itr,
  setup.pars,
  ct,
  ct.spike.times,
  individual.spike.train
)
}
\arguments{
\item{param}{a numeric vector containing the value of the eta and gamma parameters in the model.
For a model containing K frequencies, the first K numbers are eta values and the second K numbers are gamma values.
This must be the first argument as we wish to maximize the value of the log-likelihood function over the entire set of eta and gamma parameters.}

\item{f.hat}{a numeric vector containing frequency estimates for a particular model.}

\item{w0.hat.itr}{a numeric vector containing phase estimates for a particular model and spike train.}

\item{setup.pars}{a list of additional parameters for the likelihood function, computed by the \code{\link[=SetupLikelihoods]{SetupLikelihoods()}} function.}

\item{ct}{a numeric vector containing the estimated piecewise constant intensity function. The length of \code{ct} should be a whole number power of 2.}

\item{ct.spike.times}{a numeric vector containing the values of \code{ct} at the specific times a spike was recorded.}

\item{individual.spike.train}{a numeric vector containing the spike times for that spike train.}
}
\value{
The value of the log-likelihood function for the additive model.
}
\description{
Computes the log-likelihood of a spike train under the assumptions of the additive model.
}
\references{
Ramezan, R., Marriott, P., and Chenouri, S. (2014), \emph{Statistics in Medicine}, \strong{33}(2), 238-256. doi: 10.1002/sim.5923.
}
