% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrappedTheta.R
\name{bootstrapped.theta}
\alias{bootstrapped.theta}
\title{Bootstrapped Theta}
\usage{
bootstrapped.theta(theta.t, B = 1000)
}
\arguments{
\item{theta.t}{a matrix containing estimates of \theta(t). Each row represents a time point and each column reprsents a spike train.
Thus, each value in the matrix represents the value of \theta(t) at time point t estimated from a particular spike train.}

\item{B}{the number of desired bootstrap resamples}
}
\value{
a matrix containing the lower and upper bounds of the percentile confidence interval at each time point
}
\description{
Computes a 95% pointwise confidence interval for \theta(t) using the bootstrap percentile method
}
