% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctAllPoints.R
\name{ct.all.points}
\alias{ct.all.points}
\title{c(t) at all points}
\usage{
ct.all.points(t, terminal.points, ct)
}
\arguments{
\item{t}{a numeric vector of time points at which to evaluate the function}

\item{terminal.points}{a numeric vector containing the time points at which c(t) changes}

\item{ct}{a numeric vector containing the estimated piecewise constant estimated intensity function c(t). The length of ct should be a whole number power of 2.}
}
\value{
a numeric vector containing the value of c(t) at all time points at which to evaluate it
}
\description{
Computes c(t) at all arbitrary points in a time vector
}
