% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctIndividualPoint.R
\name{ct.individual.point}
\alias{ct.individual.point}
\title{c(t) at Individual Points}
\usage{
ct.individual.point(t1, terminal.points, ct)
}
\arguments{
\item{t1}{a single time value}

\item{terminal.points}{a numeric vector containing the time points at which c(t) changes}

\item{ct}{a numeric vector containing the estimated piecewise constant intensity function c(t)}
}
\value{
a single numeric value of c(t)
}
\description{
Computes the estimated piecewise constant intensity function c(t) at a particular time point t1
}
