% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CtIndividualPoint.R
\name{CtIndividualPoint}
\alias{CtIndividualPoint}
\title{\eqn{c(t)} at Individual Points}
\usage{
CtIndividualPoint(t1, terminal.points, ct)
}
\arguments{
\item{t1}{a single time value.}

\item{terminal.points}{a numeric vector containing the time points at which c(t) changes. This vector is usually of length \eqn{2^J+1}.}

\item{ct}{a numeric vector containing the estimated piecewise constant intensity function c(t).}
}
\value{
A single numeric value of c(t).
}
\description{
Computes the estimated piecewise constant intensity function c(t) at a particular time point t1. The function is defined to be right-continuous.
}
