% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatePhase.R
\name{EstimatePhase}
\alias{EstimatePhase}
\title{Estimate initial phases of the periodic function \eqn{\nu}}
\usage{
EstimatePhase(spikes, f.hat)
}
\arguments{
\item{spikes}{a list of spike trains.}

\item{f.hat}{a numeric vector containing frequency estimates for a particular model.}
}
\value{
A numeric matrix containing the estimates of the initial phases for each frequency and spike train. The rows represent individual spike trains and the columns are individual frequency values.
}
\description{
Computes the initial phases for spike trains based on the given frequencies.
}
