% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateThetaT.R
\name{EstimateThetaT}
\alias{EstimateThetaT}
\title{Estimate the intensity function \eqn{\theta(t)}}
\usage{
EstimateThetaT(
  spikes,
  f.hat.list,
  w0.hat.list,
  K.list,
  models.to.fit,
  t.start = 0,
  t.end,
  terminal.points,
  ct,
  intensity.function.length = (1 + (t.end - t.start) * 10^(3 - ceiling(log10(t.end -
    t.start))))
)
}
\arguments{
\item{spikes}{a list of spike trains.}

\item{f.hat.list}{a list containing estimated frequencies for each model.}

\item{w0.hat.list}{a list containing estimated phases for each model.}

\item{K.list}{a list of matrices containing estimated eta and gamma parameters for each model and the estimated goodness-of-fit criteria (AIC, etc.) for each model.}

\item{models.to.fit}{a list containing the names of the models to fit an intensity function for and their corresponding indices in K.list.
Typically, this list contains either all models or only the models chosen by specific goodness-of-fit criteria.}

\item{t.start}{the starting time of the recording window; the default value is 0.}

\item{t.end}{the ending time of the recording window.}

\item{terminal.points}{a numeric vector containing the time points at which the piecewise constant estimate c(t) changes.}

\item{ct}{a numeric vector containing the estimated piecewise constant intensity function.}

\item{intensity.function.length}{the number of points in the discretized intensity function.
The larger this value is, the better the resolution.
For spike trains of 10 seconds or less, the default value corresponds to 10 ms resolution.
For spike trains of 10-100 seconds, the default value corresponds to 100 ms resolution.}
}
\value{
A list of length 2.
The first item in the list is a list of matrices containing the intensity estimates (average and bootstrap CI) for each model.
The second item in the list is a list of matrices containing the intensity estimates (for each spike train) for each model.
}
\description{
Estimates average intensity function and computes bootstrap confidence intervals for a number of specified models.
}
