% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHatMinusIFn.R
\name{f.hat.minus.i.fn}
\alias{f.hat.minus.i.fn}
\title{Leave-one-out density function cross-validation}
\usage{
f.hat.minus.i.fn(til, i, points, f.hat.minus.i)
}
\arguments{
\item{til}{the time of the lth spike in trial i of a set of spike trains}

\item{i}{the index of the trial whose estimated density function is to be removed (in leave-one-out cross-validation)}

\item{points}{a numerical vector containing the end points of the dyadic partitioning for a given resolution}

\item{f.hat.minus.i}{a numeric matrix where the columns represent the m density estimates and the rows represent the time points at which the density estimates are calculated}
}
\value{
a numeric vector of the average of m-1 density function estimates for a given set of time points with the ith density estimate removed
}
\description{
Calculates the average of m-1 density function estimates for a given set of time points with the ith density estimate removed.
This function implements equation (1) of the supplementary material in Ramezan et al., 2014
}
