% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindTopFrequencies.R
\name{FindTopFrequencies}
\alias{FindTopFrequencies}
\title{Find highest-amplitude frequencies}
\usage{
FindTopFrequencies(
  spikes,
  t.start = 0,
  t.end = 10,
  freqrange = list(c(2, 30)),
  q = 5,
  default.grid.spacing = 1,
  periodogram.window.size = 25,
  default.coef.step = 0.01
)
}
\arguments{
\item{spikes}{a list of spike trains.}

\item{t.start}{the starting time of the recording window. The default value is 0.}

\item{t.end}{the ending time of the recording window. The default value is 10, corresponding to a 10-second recording.}

\item{freqrange}{a list of (non-overlapping) frequency ranges. Each item in the list should be a numeric vector of lowest and highest frequencies in the range.}

\item{q}{the number of highest-amplitude frequency components to find in each train.}

\item{default.grid.spacing}{the spacing to use in the frequency search. This can be a single number reflecting the same grid spacing over all frequency ranges or a vector of the same length as \code{freqrange}.}

\item{periodogram.window.size}{the number of points on each side of a given frequency to use when smoothing the periodogram.}

\item{default.coef.step}{the coef.step value to pass to the smoothed periodogram.}
}
\value{
A sorted table. The names of the table, in order, are the most common high-amplitude frequencies in the periodograms of the individual spike trains, and
the values in the table are the number of spike trains the frequency is of high amplitude in.
}
\description{
Finds the highest-amplitude frequency components in each spike train.
}
