% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitAdditiveModel.R
\name{FitAdditiveModel}
\alias{FitAdditiveModel}
\title{Fit additive model}
\usage{
FitAdditiveModel(spikes, f.hat, w0.hat, setup.pars, terminal.points, ct)
}
\arguments{
\item{spikes}{a list of spike trains.}

\item{f.hat}{a numeric vector containing frequency estimates for a particular model.}

\item{w0.hat}{a matrix containing phase estimates for each spike train for a particular model.}

\item{setup.pars}{a list of additional parameters for the likelihood function, computed by the \code{\link[=SetupLikelihoods]{SetupLikelihoods()}} function.}

\item{terminal.points}{a numeric vector containing the time points at which \eqn{c(t)} changes.}

\item{ct}{a numeric vector containing the estimated piecewise constant intensity function c(t). The length of ct should be a whole number power of 2.}
}
\value{
A list of length 3.
The first item in the list is a matrix whose rows each contain the MLEs of eta for a single spike train.
The second item in the list is a matrix whose rows each contain the MLEs of gamma for a single spike train.
The third item in the list is a matrix whose rows each contain the AIC, AICc, BIC, and log-likelihood for the model for a single spike train
}
\description{
Fits the additive multiscale model.
}
\references{
Ramezan, R., Marriott, P., and Chenouri, S. (2014), \emph{Statistics in Medicine}, \strong{33}(2), 238-256. doi: 10.1002/sim.5923.
}
