% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAllMultAddModels.R
\name{FitAllMultAddModels}
\alias{FitAllMultAddModels}
\title{Fit all multiplicative and additive models}
\usage{
FitAllMultAddModels(
  K,
  spikes,
  f.common.table,
  setup.pars,
  terminal.points,
  ct,
  user.select = FALSE
)
}
\arguments{
\item{K}{the number of frequency components in the largest model to fit}

\item{spikes}{a list of spike trains}

\item{f.common.table}{a table whose names contain the high-amplitude frequency components as computed by \code{FindTopFrequencies}}

\item{setup.pars}{a list of additional parameters for the likelihood function, computed by \code{SetupLikelihoods}}

\item{terminal.points}{a numeric vector containing the time points at which c(t) changes}

\item{ct}{a numeric vector containing the estimated piecewise constant intensity function c(t). The length of c(t) should be a whole number power of 2.}

\item{user.select}{whether to allow the user to select the frequencies for the model; must be FALSE for this function to run effectively}
}
\value{
a list of length 3
The first item in the list is a list of frequency estimates for each model
The second item in the list is a list of phase estimates for each model
The third item in the list is a list of eta/gamma estimates and fit criteria for each model
}
\description{
Wrapper function to fit all 2K+1 models (multiplicative, additive, and nonperiodic)
}
