% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateModelNames.R
\name{GenerateModelNames}
\alias{GenerateModelNames}
\title{Generate model names}
\usage{
GenerateModelNames(K.length)
}
\arguments{
\item{K.length}{the number of models to fit. If K is the maximum number of frequency components, then \code{K.length} = 2K+1.}
}
\value{
A character vector containing model names.
}
\description{
Creates a vector of names identifying the 2K+1 models.
}
