% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qVector.R
\name{HaslingerQ}
\alias{HaslingerQ}
\title{Quantile Vector}
\usage{
HaslingerQ(endpoints.bins, theta, resolution)
}
\arguments{
\item{endpoints.bins}{a numeric vector containing the endpoints of the bins for the binned spike train.}

\item{theta}{a numeric vector containing the average of intensity function estimates across trials.}

\item{resolution}{the time resolution of the intensity function, related to the number of points at which the function is evaluated.}
}
\value{
A numeric vector of q values from Haslinger et al. (2010).
}
\description{
Calculate the q-values for the goodness-of-fit plot.
}
\references{
Haslinger, R., Pipa G., and Brown, E. (2010). Discrete time rescaling theorem: determining goodness of fit for discrete time statistical models of neural spiking. \emph{Neural Computation}. \strong{22}(10):2477-506. doi: 10.1162/NECO_a_00015.
}
