% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyTerminalPoints.R
\name{IdentifyTerminalPoints}
\alias{IdentifyTerminalPoints}
\title{Identify terminal points}
\usage{
IdentifyTerminalPoints(t.start, t.end, J)
}
\arguments{
\item{t.start}{the starting time of the data collection.}

\item{t.end}{the ending time of the data collection.}

\item{J}{the maximum size of the tree in the initial dyadic partitioning used to estimate c(t).}
}
\value{
A numeric vector of points at which c(t) changes.
}
\description{
Computes a vector of times at which the estimated piecewise constant intensity function c(t) changes.
}
