% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISIPlot.R
\name{ISIPlot}
\alias{ISIPlot}
\title{Inter-spike interval plot}
\usage{
ISIPlot(NeuronNumber, TrialNumber, spike.train = spikes)
}
\arguments{
\item{NeuronNumber}{an identifier for the neuron being plotted}

\item{TrialNumber}{the trial number whose spike train is being plotted}

\item{spike.train}{a list of spike trains}
}
\value{
a ggplot object containing the parameters of the ISI plot
}
\description{
Produces an inter-spike interval (ISI) plot for a single trial
}
