% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu.R
\name{nu}
\alias{nu}
\title{Almost periodic function}
\usage{
nu(x)
}
\arguments{
\item{x}{an arbitrary numeric vector or scalar}
}
\value{
1 + cos(2 \pi x)
}
\description{
Computes the function \nu(x) = 1 + cos(2 \pi x)
}
