% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterFeasibilityCheck.R
\name{ParameterFeasibilityCheck}
\alias{ParameterFeasibilityCheck}
\title{Parameter check function}
\usage{
ParameterFeasibilityCheck(f.hat, w0.hat.itr, eta.hat, gama.hat, f.max = 100)
}
\arguments{
\item{f.hat}{a numeric vector containing frequency estimates for a particular model.}

\item{w0.hat.itr}{a numeric vector containing initial phase estimates for every spike train used to fit a particular model.}

\item{eta.hat}{a numeric vector containing the contribution of each periodic term to the intensity function.}

\item{gama.hat}{a numeric vector containing the amplitudes of each periodic term in the intensity function.}

\item{f.max}{a scalar indicating the absolute highest frequency of spiking activity.}
}
\value{
Warning message(s) if the constraints on parameters are violated, otherwise NULL.
}
\description{
A test function to ensure that appropriate constraints on parameters of the likelihood function are met
}
