% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonRDP.R
\name{PoissonRDP}
\alias{PoissonRDP}
\title{Poisson Recursive Dyadic Partitioning}
\source{
This code is an R translation of the \href{http://math.bu.edu/people/kolaczyk/software/msglmcode.zip}{MATLAB code} from Kolaczyk and Nowak (2005) adjusting for piecewise constant intensity function; i.e., m = 0 in the original MATLAB code.
}
\usage{
PoissonRDP(sig, gamma)
}
\arguments{
\item{sig}{a numeric vector determining the raw data, e.g., the heights (counts) of the histograms to be smoothed.}

\item{gamma}{a scalar determining the penalty factor in the penalized likelihood method.}
}
\value{
A list containing two numeric vectors:
\enumerate{
\item \code{est}: The estimated intensity function, represented as the RDP piecewise-constant estimate $c(t)$.
\item \code{splitvec}: A binary vector of length $2^J - 1$, where $2^J$ is the length of the input \code{sig}. Its entries correspond to the internal nodes of the complete dyadic partition tree, ordered breadth-first: starting with the root node, followed by nodes at successively finer scales from left to right within each level.
An entry \code{splitvec[i] = 1} indicates that the corresponding node is retained as a split, so that its two child intervals remain separate. An entry \code{splitvec[i] = 0} indicates that the node does not split.
}
}
\description{
Calculates the piecewise constant intensity function using penalized likelihood recursive dyadic partitioning method in a Poisson model.
}
\references{
Kolaczyk, E.D. and Nowak, R.D. (2004). Multiscale likelihood analysis and complexity penalized estimation, \emph{The Annals of Statistics}, \strong{32}(2), 500-527. doi: 10.1214/009053604000000076.

Kolaczyk, E.D. and Nowak, R.D. (2005). Multiscale generalized linear models for nonparametric function estimation. \emph{Biometrika},  \strong{92}(1), 119–133. doi: 10.1093/biomet/92.1.119.
}
