% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-new.R
\name{Raster}
\alias{Raster}
\title{Raster plot}
\usage{
Raster(NeuronNumber, spike.train = spikes)
}
\arguments{
\item{NeuronNumber}{the ID for the neuron to be plotted}

\item{spike.train}{a list of numeric vectors, each of which contains the spike times from a single trial of the experiment}
}
\value{
a ggplot object containing the parameters of the Raster plot
}
\description{
Produces a raster plot for a single neuron
}
