% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectBestModelByIC.R
\name{SelectBestModelByIC}
\alias{SelectBestModelByIC}
\title{Information criterion-based multiscale model selection}
\usage{
SelectBestModelByIC(IC, tolerance = 1e-06)
}
\arguments{
\item{IC}{a numeric vector of AIC, AICc, or BIC values for each model}

\item{tolerance}{a very small positive scalar accounting for machine precision and roundoff errors in the optimization.
The simplest model within tolerance of the minimum IC value is selected as the best model.}
}
\value{
The index (in the IC vector) of the selected model, unless two equally simple models could be selected, in which case NA is returned.
}
\description{
Selects the best model given a vector of AIC/AICc/BIC values for each model under consideration
}
