% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectMostCommonBestModel.R
\name{SelectMostCommonBestModel}
\alias{SelectMostCommonBestModel}
\title{Select most common best model}
\usage{
SelectMostCommonBestModel(selected.models, model.names)
}
\arguments{
\item{selected.models}{a factor vector containing names of the "best" models as chosen by some IC.}

\item{model.names}{a factor vector containing the names of all models fit to the data.}
}
\value{
The name of the overall "best" model as chosen by plurality vote over all spike trains with a clear best model.
}
\description{
Identifies the model chosen as "best" for the most spike trains.
}
