% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectTopFrequencies.R
\name{SelectTopFrequencies}
\alias{SelectTopFrequencies}
\title{Select top frequencies}
\usage{
SelectTopFrequencies(f.sorted, K, user.select = FALSE)
}
\arguments{
\item{f.sorted}{a sorted table. The names of the table, in order, are the most common frequencies in the periodograms of the individual spike trains.
This is usually output by the \code{\link[=FindTopFrequencies]{FindTopFrequencies()}} function.}

\item{K}{a scalar determining the number of periodic components to include in the model.}

\item{user.select}{whether to allow manual (user) control over accepting the identified frequencies. By default this is FALSE to allow running in batch mode.}
}
\value{
A numeric vector containing the K frequencies among the peaks of the (smoothed) periodogram which appear the most in the spike trains.
}
\description{
Identifies the K frequencies that appear most often in a given list of spike trains.
}
