% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupLikelihoods.R
\name{SetupLikelihoods}
\alias{SetupLikelihoods}
\title{Setup likelihood function}
\usage{
SetupLikelihoods(terminal.points)
}
\arguments{
\item{terminal.points}{a numeric vector containing the endpoints of the dyadic partitioning}
}
\value{
A list of length 5 is returned. The list contains the following scalars:

DeltaDi: the delta D value in the likelihood derivation; see Ramezan \emph{et al}. (2014).

Di.1: the first Di value in the likelihood derivation; see Ramezan \emph{et al}. (2014).

Di.0: the initial Di value in the likelihood derivation; see Ramezan \emph{et al}. (2014).

T.data: The length of the data recording window.

J: The resolution based on which the piecewise constant intensity function \eqn{c(t)} has been estimated.
}
\description{
Create a list of common parameters necessary to calculate likelihood functions for additive and multiplicative models.
}
\references{
Ramezan, R., Marriott, P., and Chenouri, S. (2014), \emph{Statistics in Medicine}, \strong{33}(2), 238-256. doi: 10.1002/sim.5923.
}
