% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothedPeriodogram.R
\name{SmoothedPeriodogram}
\alias{SmoothedPeriodogram}
\title{Smoothed periodogram}
\usage{
SmoothedPeriodogram(data, f, T.data, m = 5, coef.step = 0.01)
}
\arguments{
\item{data}{a numeric vector containing the spike times for the analyzed spike train.}

\item{f}{a numeric vector containing frequency estimates for a particular model.}

\item{T.data}{the length of the data recording window.}

\item{m}{a scalar determining the number of steps on each side of a given frequency over which the raw periodogram is averaged.}

\item{coef.step}{a scalar determining the resolution of frequency values (over 2m+1 of which the raw periodogram is averaged).}
}
\value{
A numerical vector of smoothed periodogram values over a range of frequencies.
}
\description{
Calculates the smoothed periodogram.
}
\references{
Brillinger, D.R. (1974), Time Series: data analysis and theory. New York, NY, Holt, Rinehart, and Winston.
}
