% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeSimulation.R
\name{SpikeSimulation}
\alias{SpikeSimulation}
\title{Spike train simulation}
\usage{
SpikeSimulation(
  nruns,
  t.start = 0,
  t.end,
  resolution = 0.001,
  intensity.function,
  pass.arg,
  envelope.function = function(t) return(1)
)
}
\arguments{
\item{nruns}{a scalar determining the number of spike trains (trials) to be simulated.}

\item{t.start}{the starting time of the simulated data collection.}

\item{t.end}{the ending time of the simulated data collection.}

\item{resolution}{the time resolution of the spike train. The default is 0.001 (1 ms).}

\item{intensity.function}{any intensity function for the corresponding process from which you want to simulate data.}

\item{pass.arg}{a list of arguments to pass to the intensity function. In the additive and multiplicative models, this contains four numeric vectors;
the first vector determines the frequencies, the second vector determines initial phases, the third vector determines the \eqn{\eta} values and the fourth vector determines the \eqn{\gamma} values.}

\item{envelope.function}{The envelope function used for acceptance-rejection sampling to simulate the data.}
}
\value{
A list of numeric vectors, each of which contains a simulated spike train.
}
\description{
Simulates spike trains from an inhomogeneous Poisson process with a given time-varying intensity function of one of the forms discussed in Ramezan \emph{et al.} (2014).
}
\references{
Ramezan, R., Marriott, P., and Chenouri, S. (2014), \emph{Statistics in Medicine}, \strong{33}(2), 238-256. doi: 10.1002/sim.5923.
}
