% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThetaMultiplicative.R
\name{ThetaMultiplicative}
\alias{ThetaMultiplicative}
\title{Multiplicative intensity function \eqn{\theta(t)}}
\usage{
ThetaMultiplicative(t, f, w0, eta, gamma, terminal.points, ct)
}
\arguments{
\item{t}{a numeric vector of time points at which to evaluate the function.}

\item{f}{a numeric vector containing frequency values.}

\item{w0}{a numeric vector containing initial phase values.}

\item{eta}{a numeric vector containing \eqn{\eta} values (contribution of each periodic component to the intensity function).}

\item{gamma}{a numeric vector containing \eqn{\gamma} values (amplitude of each periodic component in the function).}

\item{terminal.points}{a numeric vector containing the endpoints of the dyadic partitioning.}

\item{ct}{a numeric vector containing the estimated piecewise constant intensity function \eqn{c(t)}. The length of ct should be a whole number power of 2.}
}
\value{
A numeric vector containing the values of the multiplicative intensity function calculated at given time points.
}
\description{
Calculates the multiplicative intensity function \eqn{\theta(t)} introduced in Ramezan \emph{et al.} (2014).
}
\references{
Ramezan, R., Marriott, P., and Chenouri, S. (2014), \emph{Statistics in Medicine}, \strong{33}(2), 238-256. doi: 10.1002/sim.5923.
}
