% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaFunction.R
\name{ThetaT}
\alias{ThetaT}
\title{Intensity function \eqn{\theta(t)}}
\usage{
ThetaT(t, theta, t.start = 0, resolution)
}
\arguments{
\item{t}{the time at which to calculate \eqn{\theta(t)}. The function \code{\link{HaslingerQ}} integrates this function, so t can be of any length (including 1).}

\item{theta}{a numeric vector containing the estimated intensity function \eqn{\theta(t)}.}

\item{t.start}{the starting time of the recording}

\item{resolution}{the time difference between successive points in the theta vector}
}
\value{
The value of the intensity function \eqn{\theta(t)} at (a) specific time point(s)
}
\description{
Calculates the value of \eqn{\theta(t)} at an individual time point t
}
