% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCt.R
\name{FindCt}
\alias{FindCt}
\title{Find \eqn{c(t)}}
\usage{
FindCt(spikes, t.start, t.end, lambda, J, PSTH = TRUE)
}
\arguments{
\item{spikes}{a list of spike trains.}

\item{t.start}{the starting time of the recording window; the default value is 0.}

\item{t.end}{the ending time of the recording window.}

\item{lambda}{a penalty term used in estimating the piecewise constant intensity function c(t). Larger values of lambda result in "smoother" estimates.}

\item{J}{the maximum size of the tree in the initial dyadic partitioning used to estimate c(t). The final estimate of c(t) will have 2^J values.}

\item{PSTH}{if TRUE, will aggregate the spikes across all trials in each and every bin and then estimate c(t) using the post-stimulus-time histogram.
If FALSE, will estimate a c(t) for each trial and average the c(t) estimates.}
}
\value{
A list of length 2 is returned.
The first item in the list is the estimate of \eqn{c(t)} created by averaging the estimates for each spike train.
The second item is a matrix in which each row represents the estimate of c(t) for an individual spike train.
}
\description{
Compute the piecewise constant estimate of the intensity function, \eqn{c(t)}, for each spike train
}
