% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qVector.R
\name{HaslingerQ}
\alias{HaslingerQ}
\title{Quantile Vector}
\usage{
HaslingerQ(endpoints.bins, theta, resolution)
}
\arguments{
\item{endpoints.bins}{a numeric vector containing the endpoints of the bins for the binned spike train}

\item{theta}{a numeric vector containing the the average of intensity function estimates across trials}

\item{resolution}{the time resolution of the intensity function, related to the number of points at which the function is evaluated}
}
\value{
a numeric vector of q values from Haslinger (2010)
}
\description{
Calculate the q-values for the goodness-of-fit plot
}
