% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateYi.R
\name{HaslingerYi}
\alias{HaslingerYi}
\title{Calculate y(i)}
\usage{
HaslingerYi(spike.train, endpoints.bins, q)
}
\arguments{
\item{spike.train}{a list of numeric vectors, each of which contains the spike times for one trial of the experiment.}

\item{endpoints.bins}{a numeric vector containing the endpoints of the bins for the binned spike train.}

\item{q}{a numeric vector containing the q values calculated by the \code{\link{HaslingerQ}} function.}
}
\value{
A numeric vector of Yi values as defined in equation (39) of Haslinger et al. (2010).
}
\description{
Calculates Yi in equation (39) of Haslinger et al. (2010)
}
\references{
Haslinger, R., Pipa G., and Brown, E. (2010). Discrete time rescaling theorem: determining goodness of fit for discrete time statistical models of neural spiking. \emph{Neural Computation}. \strong{22}(10):2477-506. doi: 10.1162/NECO_a_00015.
}
