% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISIPlot.R
\name{ISIPlot}
\alias{ISIPlot}
\title{Inter-spike interval plot}
\usage{
ISIPlot(NeuronNumber, TrialNumber, spike.train)
}
\arguments{
\item{NeuronNumber}{an identifier for the neuron being plotted.}

\item{TrialNumber}{the trial number whose spike train is being plotted.}

\item{spike.train}{a list of spike trains.}
}
\value{
A ggplot object containing the parameters of the ISI plot.
}
\description{
Produces an inter-spike interval (ISI) plot for a single trial.
}
\references{
Ramezan, R., Marriott, P., and Chenouri, S. (2014), \emph{Statistics in Medicine}, \strong{33}(2), 238-256. doi: 10.1002/sim.5923.
}
