% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu.R
\name{nu}
\alias{nu}
\title{A periodic function}
\usage{
nu(x)
}
\arguments{
\item{x}{an arbitrary numeric vector or scalar.}
}
\value{
The value of the function \eqn{1 + cos(2 \pi x)}.
}
\description{
Computes the function \eqn{\nu(x) = 1 + cos(2 \pi x)}.
}
