% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonRDP.R
\name{PoissonRDP}
\alias{PoissonRDP}
\title{Poisson Recursive Dyadic Partitioning}
\source{
This code is based on \url{MATLAB code}{http://math.bu.edu/people/kolaczyk/software/msglmcode.zip} from Kolaczyk and Nowak (2005). Drs. Kolaczyk and Nowak have agreed to allow our translation in the package.
}
\usage{
PoissonRDP(sig, gamma)
}
\arguments{
\item{sig}{a numeric vector determining the raw data, e.g., the heights (counts) of the histograms to be smoothed}

\item{gamma}{a scalar determining the penalty factor in the penalized likelihood method}
}
\value{
a numeric vector of estimated c(t), the smoothed histogram of data.
}
\description{
Penalized recursive dyadic partitioning based on a Poisson model.
}
\details{
Calculates the piecewise constant intensity function using penalized likelihood  and recursive dyadic partitioning methods.
}
\references{
Kolaczyk, E.D. and Nowak, R.D. (2004). Multiscale likelihood analysis and complexity penalized estimation, \emph{The Annals of Statistics}, \strong{32}(2), 500-527. doi: 10.1214/009053604000000076.
Kolaczyk, E.D. and Nowak, R.D. (2005). Multiscale generalized linear models for nonparametric function estimation. \emph{Biometrika},  \strong{92}(1), 119–133. doi: 10.1093/biomet/92.1.119.
}
