% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasterPlot.R
\name{RasterPlot}
\alias{RasterPlot}
\title{Raster plot}
\usage{
RasterPlot(NeuronNumber, spike.train)
}
\arguments{
\item{NeuronNumber}{the ID for the neuron to be plotted (only used in the title of the raster plot).}

\item{spike.train}{a list of numeric vectors, each of which contains the spike times from a single trial of the experiment.}
}
\value{
A ggplot object containing the parameters of the Raster plot.
}
\description{
Produces a raster plot for a single neuron.
}
