% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaFunction.R
\name{theta.function}
\alias{theta.function}
\title{\theta(t)}
\usage{
theta.function(t, theta, t.start = 0, resolution)
}
\arguments{
\item{t}{the time at which to calculate \theta(t). The function \code{q.vector} integrates this function, so t can be of any length (including 1).}

\item{theta}{a numeric vector containing the estimated intensity function \theta(t).}

\item{t.start}{the starting time of the recording}

\item{resolution}{the time difference between successive points in the theta vector}
}
\value{
the value of the intensity function \theta(t) at a specific time point
}
\description{
Calculate the value of theta(t) at an individual time point t
}
